Renamed files and added Ortho and Persp buttons in the Settings.vue
