+++ src/store/index.ts
@@ -2,7 +2,7 @@ import Vue from 'vue';
 import Vuex from 'vuex';
 import counter from './modules/counter';
 import cubestore from './modules/cubestore';
-import inputslider from './modules/inputslider';
+import settings from './modules/settings';
 Vue.use(Vuex);
 
 const debug = process.env.NODE_ENV !== 'production';
@@ -11,7 +11,7 @@ export default new Vuex.Store({
     modules: {
         counter,
         cubestore,
-        inputslider,
+        settings,
     },
     strict: debug,
 });
